1. "Symmetric difference" methods renamed: * `Set::symmetricDifference()` => `Set::symmetricDifferenceNonStrict()`; * `Set::symmetricDifferenceStrict()` => `Set::symmetricDifference()`. 2. Tests renamed for these methods.
