+++ tests/Set/SymmetricDifferenceTest.php
@@ -60,131 +60,88 @@ class SymmetricDifferenceTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    [null, null],
+                    [1, 2, 3, 4, 5],
+                    [1, 2, 3, 4, 5],
                 ],
                 [],
-            ],
-            [
-                [
-                    [null, null],
-                    [],
-                ],
-                [null, null],
-            ],
-            [
-                [
-                    [null, null, null],
-                    [null, null],
-                    [null],
-                ],
-                [null],
-            ],
-            [
-                [
-                    [1, 2, 3],
-                    [2, 3, 4],
-                ],
-                [1, 4],
             ],
             [
                 [
                     [1, 2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6, 7],
+                    [1, 2, 3, '4', '5'],
                 ],
-                [1, 7],
+                [4, '4', 5, '5'],
             ],
             [
                 [
-                    [2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6, 7],
+                    ['1', '2', '3', 4, 5],
+                    [1, 2, 3, '4', '5'],
                 ],
-                [7],
+                ['1', 1, '2', 2, '3', 3, 4, '4', 5, '5'],
             ],
             [
                 [
-                    [1, 2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6],
+                    [1, 1, 2, 2, 1, 1],
+                    [2, 2, 1, 1, 2, 2],
                 ],
-                [1],
+                [1, 1, 2, 2],
             ],
             [
                 [
-                    [1, 1, 1, 1, 1, 3],
-                    [2, 2, 3, 1, 1],
-                    [3, 1, 1, 3, 3],
+                    [1, 1, 2, 2, 1, 1],
+                    [2, 2, 1, 1, '2', '2'],
                 ],
-                [1, 1, 1, 2, 2, 3, 3],
+                [1, 1, '2', '2'],
             ],
             [
                 [
-                    [1, 2, 3, 4, 5],
-                    [2, 3, 4, 5, 6, 1],
-                    [3, 4, 5, 6, 1],
+                    [1, 1, 2, 2, 1, 1],
+                    [2, 2, '1', '1', 2, 2],
                 ],
-                [],
+                [1, 1, 1, 1, 2, 2, '1', '1'],
             ],
             [
                 [
-                    [1, 2, 3, 4, 3, 5],
-                    [2, 3, 4, 5, 6, 1, 2, 3],
-                    [3, 4, 5, 6, 1, 2, 3],
+                    [1, 1, 1, 1, 1],
+                    [1, 2, 3, 4, 5, 1, 2, 3, 4, 5],
+                    [5, 5, 5, 5, 5, 1, 5, 5, 1],
                 ],
-                [2],
+                [1, 1, 1, 5, 5, 5, 5, 5, 2, 2, 3, 3, 4, 4],
             ],
             [
                 [
-                    [2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6],
+                    [1, 1, '1', '1'],
+                    [1, 2, 3, 1, 2, 3],
+                    ['2', '3'],
                 ],
-                [],
+                ['2', 2, 2, '3', '1', '1', 3, 3],
             ],
             [
                 [
-                    [2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [6],
+                    [null, 1, 6, null],
+                    [null, 2, 7, 11, null],
+                    [null, 3, 8, 12, 15, null],
+                    [null, 4, 9, 13, 16, 18, null],
+                    [null, 5, 10, 14, 17, 19, 20, null],
                 ],
-                [],
+                [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
-                    [1.1, 2.1, 3.1],
-                    [2.1, 3.1, 4.1],
+                    [null, 1, 6, null],
+                    [null, '1', 7, 11, null],
+                    [null, '1.0', 8, 12, 15, null],
+                    [null, true, 9, 13, 16, 18, null],
+                    [null, 'true', 10, 14, 17, 19, 20, null],
                 ],
-                [1.1, 4.1],
-            ],
-            [
-                [
-                    ['1', '2', '3'],
-                    ['2', '3', '4'],
-                ],
-                ['1', '4'],
-            ],
-            [
-                [
-                    ['1', '2', '3', '2', '11'],
-                    ['2', '3', '4', '12', '13', '2'],
-                ],
-                ['1', '4', '12', '11', '13'],
+                [1, '1', '1.0', true, 'true', 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
                     ['1', 2, '3.3', true, false],
-                    [true, '2', 3.3, '4', true],
+                    [true, '2', 3.3, '4', '5'],
                 ],
-                [4, false],
-            ],
-            [
-                [
-                    ['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript'],
-                    ['php', 'python', 'javascript', 'typescript'],
-                    ['php', 'java', 'c#', 'typescript'],
-                ],
-                ['c++'],
+                ['1', 2, '2', '3.3', 3.3, '4', false, '5'],
             ],
         ];
     }
@@ -242,131 +199,88 @@ class SymmetricDifferenceTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $gen([null, null]),
+                    $gen([1, 2, 3, 4, 5]),
+                    $gen([1, 2, 3, 4, 5]),
                 ],
                 [],
-            ],
-            [
-                [
-                    $gen([null, null]),
-                    $gen([]),
-                ],
-                [null, null],
-            ],
-            [
-                [
-                    $gen([null, null, null]),
-                    $gen([null, null]),
-                    $gen([null]),
-                ],
-                [null],
-            ],
-            [
-                [
-                    $gen([1, 2, 3]),
-                    $gen([2, 3, 4]),
-                ],
-                [1, 4],
             ],
             [
                 [
                     $gen([1, 2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6, 7]),
+                    $gen([1, 2, 3, '4', '5']),
                 ],
-                [1, 7],
+                [4, '4', 5, '5'],
             ],
             [
                 [
-                    $gen([2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6, 7]),
+                    $gen(['1', '2', '3', 4, 5]),
+                    $gen([1, 2, 3, '4', '5']),
                 ],
-                [7],
+                ['1', 1, '2', 2, '3', 3, 4, '4', 5, '5'],
             ],
             [
                 [
-                    $gen([1, 2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6]),
+                    $gen([1, 1, 2, 2, 1, 1]),
+                    $gen([2, 2, 1, 1, 2, 2]),
                 ],
-                [1],
+                [1, 1, 2, 2],
             ],
             [
                 [
-                    $gen([1, 1, 1, 1, 1, 3]),
-                    $gen([2, 2, 3, 1, 1]),
-                    $gen([3, 1, 1, 3, 3]),
+                    $gen([1, 1, 2, 2, 1, 1]),
+                    $gen([2, 2, 1, 1, '2', '2']),
                 ],
-                [1, 1, 1, 2, 2, 3, 3],
+                [1, 1, '2', '2'],
             ],
             [
                 [
-                    $gen([1, 2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6, 1]),
-                    $gen([3, 4, 5, 6, 1]),
+                    $gen([1, 1, 2, 2, 1, 1]),
+                    $gen([2, 2, '1', '1', 2, 2]),
                 ],
-                [],
+                [1, 1, 1, 1, 2, 2, '1', '1'],
             ],
             [
                 [
-                    $gen([1, 2, 3, 4, 3, 5]),
-                    $gen([2, 3, 4, 5, 6, 1, 2, 3]),
-                    $gen([3, 4, 5, 6, 1, 2, 3]),
+                    $gen([1, 1, 1, 1, 1]),
+                    $gen([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $gen([5, 5, 5, 5, 5, 1, 5, 5, 1]),
                 ],
-                [2],
+                [1, 1, 1, 5, 5, 5, 5, 5, 2, 2, 3, 3, 4, 4],
             ],
             [
                 [
-                    $gen([2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6]),
+                    $gen([1, 1, '1', '1']),
+                    $gen([1, 2, 3, 1, 2, 3]),
+                    $gen(['2', '3']),
                 ],
-                [],
+                ['2', 2, 2, '3', '1', '1', 3, 3],
             ],
             [
                 [
-                    $gen([2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([6]),
+                    $gen([null, 1, 6, null]),
+                    $gen([null, 2, 7, 11, null]),
+                    $gen([null, 3, 8, 12, 15, null]),
+                    $gen([null, 4, 9, 13, 16, 18, null]),
+                    $gen([null, 5, 10, 14, 17, 19, 20, null]),
                 ],
-                [],
+                [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
-                    $gen([1.1, 2.1, 3.1]),
-                    $gen([2.1, 3.1, 4.1]),
+                    $gen([null, 1, 6, null]),
+                    $gen([null, '1', 7, 11, null]),
+                    $gen([null, '1.0', 8, 12, 15, null]),
+                    $gen([null, true, 9, 13, 16, 18, null]),
+                    $gen([null, 'true', 10, 14, 17, 19, 20, null]),
                 ],
-                [1.1, 4.1],
-            ],
-            [
-                [
-                    $gen(['1', '2', '3']),
-                    $gen(['2', '3', '4']),
-                ],
-                ['1', '4'],
-            ],
-            [
-                [
-                    $gen(['1', '2', '3', '2', '11']),
-                    $gen(['2', '3', '4', '12', '13', '2']),
-                ],
-                ['1', '4', '12', '11', '13'],
+                [1, '1', '1.0', true, 'true', 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
                     $gen(['1', 2, '3.3', true, false]),
-                    $gen([true, '2', 3.3, '4', true]),
+                    $gen([true, '2', 3.3, '4', '5']),
                 ],
-                [4, false],
-            ],
-            [
-                [
-                    $gen(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $gen(['php', 'python', 'javascript', 'typescript']),
-                    $gen(['php', 'java', 'c#', 'typescript']),
-                ],
-                ['c++'],
+                ['1', 2, '2', '3.3', 3.3, '4', false, '5'],
             ],
         ];
     }
@@ -424,131 +338,88 @@ class SymmetricDifferenceTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $iter([null, null]),
+                    $iter([1, 2, 3, 4, 5]),
+                    $iter([1, 2, 3, 4, 5]),
                 ],
                 [],
-            ],
-            [
-                [
-                    $iter([null, null]),
-                    $iter([]),
-                ],
-                [null, null],
-            ],
-            [
-                [
-                    $iter([null, null, null]),
-                    $iter([null, null]),
-                    $iter([null]),
-                ],
-                [null],
-            ],
-            [
-                [
-                    $iter([1, 2, 3]),
-                    $iter([2, 3, 4]),
-                ],
-                [1, 4],
             ],
             [
                 [
                     $iter([1, 2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7]),
+                    $iter([1, 2, 3, '4', '5']),
                 ],
-                [1, 7],
+                [4, '4', 5, '5'],
             ],
             [
                 [
-                    $iter([2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7]),
+                    $iter(['1', '2', '3', 4, 5]),
+                    $iter([1, 2, 3, '4', '5']),
                 ],
-                [7],
+                ['1', 1, '2', 2, '3', 3, 4, '4', 5, '5'],
             ],
             [
                 [
-                    $iter([1, 2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6]),
+                    $iter([1, 1, 2, 2, 1, 1]),
+                    $iter([2, 2, 1, 1, 2, 2]),
                 ],
-                [1],
+                [1, 1, 2, 2],
             ],
             [
                 [
-                    $iter([1, 1, 1, 1, 1, 3]),
-                    $iter([2, 2, 3, 1, 1]),
-                    $iter([3, 1, 1, 3, 3]),
+                    $iter([1, 1, 2, 2, 1, 1]),
+                    $iter([2, 2, 1, 1, '2', '2']),
                 ],
-                [1, 1, 1, 2, 2, 3, 3],
+                [1, 1, '2', '2'],
             ],
             [
                 [
-                    $iter([1, 2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6, 1]),
-                    $iter([3, 4, 5, 6, 1]),
+                    $iter([1, 1, 2, 2, 1, 1]),
+                    $iter([2, 2, '1', '1', 2, 2]),
                 ],
-                [],
+                [1, 1, 1, 1, 2, 2, '1', '1'],
             ],
             [
                 [
-                    $iter([1, 2, 3, 4, 3, 5]),
-                    $iter([2, 3, 4, 5, 6, 1, 2, 3]),
-                    $iter([3, 4, 5, 6, 1, 2, 3]),
+                    $iter([1, 1, 1, 1, 1]),
+                    $iter([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $iter([5, 5, 5, 5, 5, 1, 5, 5, 1]),
                 ],
-                [2],
+                [1, 1, 1, 5, 5, 5, 5, 5, 2, 2, 3, 3, 4, 4],
             ],
             [
                 [
-                    $iter([2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6]),
+                    $iter([1, 1, '1', '1']),
+                    $iter([1, 2, 3, 1, 2, 3]),
+                    $iter(['2', '3']),
                 ],
-                [],
+                ['2', 2, 2, '3', '1', '1', 3, 3],
             ],
             [
                 [
-                    $iter([2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([6]),
+                    $iter([null, 1, 6, null]),
+                    $iter([null, 2, 7, 11, null]),
+                    $iter([null, 3, 8, 12, 15, null]),
+                    $iter([null, 4, 9, 13, 16, 18, null]),
+                    $iter([null, 5, 10, 14, 17, 19, 20, null]),
                 ],
-                [],
+                [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
-                    $iter([1.1, 2.1, 3.1]),
-                    $iter([2.1, 3.1, 4.1]),
+                    $iter([null, 1, 6, null]),
+                    $iter([null, '1', 7, 11, null]),
+                    $iter([null, '1.0', 8, 12, 15, null]),
+                    $iter([null, true, 9, 13, 16, 18, null]),
+                    $iter([null, 'true', 10, 14, 17, 19, 20, null]),
                 ],
-                [1.1, 4.1],
-            ],
-            [
-                [
-                    $iter(['1', '2', '3']),
-                    $iter(['2', '3', '4']),
-                ],
-                ['1', '4'],
-            ],
-            [
-                [
-                    $iter(['1', '2', '3', '2', '11']),
-                    $iter(['2', '3', '4', '12', '13', '2']),
-                ],
-                ['1', '4', '12', '11', '13'],
+                [1, '1', '1.0', true, 'true', 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
                     $iter(['1', 2, '3.3', true, false]),
-                    $iter([true, '2', 3.3, '4', true]),
+                    $iter([true, '2', 3.3, '4', '5']),
                 ],
-                [4, false],
-            ],
-            [
-                [
-                    $iter(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $iter(['php', 'python', 'javascript', 'typescript']),
-                    $iter(['php', 'java', 'c#', 'typescript']),
-                ],
-                ['c++'],
+                ['1', 2, '2', '3.3', 3.3, '4', false, '5'],
             ],
         ];
     }
@@ -606,138 +477,95 @@ class SymmetricDifferenceTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $trav([null, null]),
+                    $trav([1, 2, 3, 4, 5]),
+                    $trav([1, 2, 3, 4, 5]),
                 ],
                 [],
-            ],
-            [
-                [
-                    $trav([null, null]),
-                    $trav([]),
-                ],
-                [null, null],
-            ],
-            [
-                [
-                    $trav([null, null, null]),
-                    $trav([null, null]),
-                    $trav([null]),
-                ],
-                [null],
-            ],
-            [
-                [
-                    $trav([1, 2, 3]),
-                    $trav([2, 3, 4]),
-                ],
-                [1, 4],
             ],
             [
                 [
                     $trav([1, 2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6, 7]),
+                    $trav([1, 2, 3, '4', '5']),
                 ],
-                [1, 7],
+                [4, '4', 5, '5'],
             ],
             [
                 [
-                    $trav([2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6, 7]),
+                    $trav(['1', '2', '3', 4, 5]),
+                    $trav([1, 2, 3, '4', '5']),
                 ],
-                [7],
+                ['1', 1, '2', 2, '3', 3, 4, '4', 5, '5'],
             ],
             [
                 [
-                    $trav([1, 2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6]),
+                    $trav([1, 1, 2, 2, 1, 1]),
+                    $trav([2, 2, 1, 1, 2, 2]),
                 ],
-                [1],
+                [1, 1, 2, 2],
             ],
             [
                 [
-                    $trav([1, 1, 1, 1, 1, 3]),
-                    $trav([2, 2, 3, 1, 1]),
-                    $trav([3, 1, 1, 3, 3]),
+                    $trav([1, 1, 2, 2, 1, 1]),
+                    $trav([2, 2, 1, 1, '2', '2']),
                 ],
-                [1, 1, 1, 2, 2, 3, 3],
+                [1, 1, '2', '2'],
             ],
             [
                 [
-                    $trav([1, 2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6, 1]),
-                    $trav([3, 4, 5, 6, 1]),
+                    $trav([1, 1, 2, 2, 1, 1]),
+                    $trav([2, 2, '1', '1', 2, 2]),
                 ],
-                [],
+                [1, 1, 1, 1, 2, 2, '1', '1'],
             ],
             [
                 [
-                    $trav([1, 2, 3, 4, 3, 5]),
-                    $trav([2, 3, 4, 5, 6, 1, 2, 3]),
-                    $trav([3, 4, 5, 6, 1, 2, 3]),
+                    $trav([1, 1, 1, 1, 1]),
+                    $trav([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $trav([5, 5, 5, 5, 5, 1, 5, 5, 1]),
                 ],
-                [2],
+                [1, 1, 1, 5, 5, 5, 5, 5, 2, 2, 3, 3, 4, 4],
             ],
             [
                 [
-                    $trav([2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6]),
+                    $trav([1, 1, '1', '1']),
+                    $trav([1, 2, 3, 1, 2, 3]),
+                    $trav(['2', '3']),
                 ],
-                [],
+                ['2', 2, 2, '3', '1', '1', 3, 3],
             ],
             [
                 [
-                    $trav([2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([6]),
+                    $trav([null, 1, 6, null]),
+                    $trav([null, 2, 7, 11, null]),
+                    $trav([null, 3, 8, 12, 15, null]),
+                    $trav([null, 4, 9, 13, 16, 18, null]),
+                    $trav([null, 5, 10, 14, 17, 19, 20, null]),
                 ],
-                [],
+                [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
-                    $trav([1.1, 2.1, 3.1]),
-                    $trav([2.1, 3.1, 4.1]),
+                    $trav([null, 1, 6, null]),
+                    $trav([null, '1', 7, 11, null]),
+                    $trav([null, '1.0', 8, 12, 15, null]),
+                    $trav([null, true, 9, 13, 16, 18, null]),
+                    $trav([null, 'true', 10, 14, 17, 19, 20, null]),
                 ],
-                [1.1, 4.1],
-            ],
-            [
-                [
-                    $trav(['1', '2', '3']),
-                    $trav(['2', '3', '4']),
-                ],
-                ['1', '4'],
-            ],
-            [
-                [
-                    $trav(['1', '2', '3', '2', '11']),
-                    $trav(['2', '3', '4', '12', '13', '2']),
-                ],
-                ['1', '4', '12', '11', '13'],
+                [1, '1', '1.0', true, 'true', 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
             ],
             [
                 [
                     $trav(['1', 2, '3.3', true, false]),
-                    $trav([true, '2', 3.3, '4', true]),
+                    $trav([true, '2', 3.3, '4', '5']),
                 ],
-                [4, false],
-            ],
-            [
-                [
-                    $trav(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $trav(['php', 'python', 'javascript', 'typescript']),
-                    $trav(['php', 'java', 'c#', 'typescript']),
-                ],
-                ['c++'],
+                ['1', 2, '2', '3.3', 3.3, '4', false, '5'],
             ],
         ];
     }
 
     /**
      * @dataProvider dataProviderForMixed
-     * @param        array<\iterable> $iterables
+     * @param        array<mixed> $iterables
      * @param        array $expected
      */
     public function testMixed(array $iterables, array $expected): void
@@ -780,48 +608,30 @@ class SymmetricDifferenceTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    [null, null, null, null, null],
-                    $gen([null, null, null, null]),
-                    $iter([null, null, null]),
-                    $trav([null, null]),
+                    [1, 1, 1],
+                    $gen([1, 2, 3, 1, 2, 3]),
+                    $iter([3, 3, 3, 1]),
+                    $trav([]),
                 ],
-                [null],
+                [1, 3, 2, 2],
             ],
             [
                 [
-                    [null, null, null, null, null, null],
-                    $gen([null, null, null, null]),
-                    $iter([null, null, null]),
-                    $trav([null, null]),
+                    [1, 1, 1],
+                    $gen([1, 2, 3, 1, 2, 3]),
+                    $iter(['3', 3, 3, 1]),
+                    $trav([]),
                 ],
-                [null, null],
+                [1, '3', 2, 2],
             ],
             [
                 [
-                    [1, 2, 3, 4, 5],
-                    $gen([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7]),
-                    $trav([4, 5, 6, 7, 8]),
+                    [1, 1, 1, 1, 1],
+                    $gen([1, 1, 1, 1, '1']),
+                    $iter([1, 1, 1, 1, 1.0]),
+                    $trav([1, 1, 1, 1, true]),
                 ],
-                [1, 8],
-            ],
-            [
-                [
-                    [1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5],
-                    $gen([2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7, 3, 4, 5, 6, 7, 3, 4, 5, 6, 7]),
-                    $trav([4, 5, 6, 7, 8, 4, 5, 6, 7, 8, 4, 5, 6, 7, 8]),
-                ],
-                [1, 1, 1, 8, 8, 8],
-            ],
-            [
-                [
-                    [1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 11],
-                    $gen([2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 12, 15]),
-                    $iter([3, 4, 5, 6, 7, 3, 4, 5, 6, 7, 3, 4, 5, 6, 7, 13, 16, 18]),
-                    $trav([4, 5, 6, 7, 8, 4, 5, 6, 7, 8, 4, 5, 6, 7, 8, 14, 17, 19, 20]),
-                ],
-                [1, 1, 1, 8, 8, 8, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
+                [1, '1', 1.0, true],
             ],
         ];
     }
